# public/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "ProjectName"
host: node-express-joi-starter.herokuapp.com
basePath: /
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
definitions:
  user:
    type: object
    properties:
      active:
        type: boolean
      online:
        type: boolean
      username:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      email:
        type: string
      phone:
        type: string
      pin:
        type: number
      lastLogin:
        type: string
        format: date-time
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time

paths:
  /api/user:
    post:
      summary: Create new User
      tags:
        - user
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            properties:
              username:
                type: string
                required: true
              email:
                type: string
              phone:
                type: string
              password:
                type: string
                required: true
      responses:
        '201':
          description: Created User
          schema:
            $ref: '#/definitions/user'
        '400':
          description: Bad request
  /api/user/{id}:
    get:
      summary: Get User By Id
      tags:
        - user/:id
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
          description: user Id
          required: true
      responses:
        '200':
          description: User
          schema:
            $ref: '#/definitions/user'
        '400':
          description: Bad request
        '404':
          description: Not Found
    patch:
      summary: Update User by Id
      tags:
        - user/:id
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
          description: user Id
          required: true
        - name: body
          in: body
          schema:
            $ref: '#/definitions/user'
      responses:
        '200':
          description: Updated User
          schema:
            $ref: '#/definitions/user'
        '400':
          description: Bad request
        '404':
          description: Not Found
    delete:
      summary: Soft Delete a User by Id
      tags:
        - user/:id
      parameters:
        - name: id
          in: path
          type: string
          description: User Id
          required: true
      responses:
        '200':
          description: Return object
          schema:
            type: object
            properties:
              ok:
                type: integer
                enum: [0, 1]
              nModified:
                type: integer
                enum: [0, 1]
              n:
                type: integer
                enum: [0, 1]
        '400':
          description: Bad request
        '404':
          description: Not Found
  /api/auth:
    post:
      summary: Auth user by username and password
      tags:
        - auth
      parameters:
        - name: username
          in: query
          type: string
          required: true
        - name: password
          in: query
          type: string
          required: true
      responses:
        '200':
          description: Return object
          schema:
            type: object
            properties:
              token:
                type: string
        '400':
          description: Bad request
  /healthcheck:
    get:
      summary: API Healthcheck
      tags:
        - healthcheck
      parameters:
      responses:
        '200':
          description: Return object
          schema:
            type: string